backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
SQLAlchemy==2.0.30
aiomysql==0.2.0
PyMySQL==1.1.1
alembic==1.13.1
openai==1.30.0
tiktoken==0.7.0
redis==5.0.4
hiredis==2.3.2
pydantic==2.7.1
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
loguru==0.7.2
pytest==8.2.1
pytest-asyncio==0.23.7
httpx==0.27.0
pytest-cov==5.0.0
python-dotenv==1.0.1